organize module `lineup` by hiding the `MatchLineupClient` // splitting up the `MatchLineupClient` to fetch match-ids and fetch specific match

+++ src/main/kotlin/org/unividuell/news/comunio/match/MatchAiController.kt
@@ -29,7 +29,7 @@ class MatchAiController(
     @GetMapping("/ai/matches/{groupOrderId}/{matchId}", produces = ["text/html"])
     fun match(
         @PathVariable groupOrderId: Int,
-        @PathVariable matchId: Long,
+        @PathVariable matchId: Int,
     ): String? {
         val matches = matchComposer
             .composeMatch(groupOrderId = groupOrderId)
